requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pyyaml>=6.0
httpx>=0.24.0
